+++ lib/applitools/selenium/browsers_info.rb
@@ -4,7 +4,7 @@ module Applitools
   module Selenium
     class BrowsersInfo < Set
       def add(obj)
-        return super if obj.is_a? Applitools::Selenium::RenderBrowserInfo
+        return super if obj.is_a? Applitools::Selenium::IRenderBrowserInfo
         raise(
           Applitools::EyesIllegalArgument,
           'It is expected the value to be an Applitools::Selenium::RenderBrowserInfo instance,' \
@@ -15,7 +15,7 @@ module Applitools
       def each(viewport_size = nil)
         return super() unless empty?
         return unless viewport_size
-        default = Applitools::Selenium::RenderBrowserInfo.new.tap do |bi|
+        default = Applitools::Selenium::DesktopBrowserInfo.new.tap do |bi|
           bi.viewport_size = viewport_size
           bi.browser_type = BrowserTypes::CHROME
         end