+++ lib/applitools/selenium/visual_grid/ios_device_name.rb
@@ -1,0 +1,30 @@
+module IosDeviceName
+  extend self
+  IPhone_11_Pro = 'iPhone 11 Pro'
+  IPhone_11_Pro_Max = 'iPhone 11 Pro Max'
+  IPhone_11 = 'iPhone 11'
+  IPhone_XR = 'iPhone XR'
+  IPhone_XS =  'iPhone Xs'
+  IPhone_X = 'iPhone X'
+  IPhone_8 = 'iPhone 8'
+  IPhone_7 = 'iPhone 7'
+  IPad_Pro_3 = 'iPad Pro (12.9-inch) (3rd generation)'
+  IPad_7 = 'iPad (7th generation)'
+  IPad_Air_2 = 'iPad Air (2nd generation)'
+
+  def enum_values
+    [
+      IPhone_11_Pro,
+      IPhone_11_Pro_Max,
+      IPhone_11,
+      IPhone_XR,
+      IPhone_XS,
+      IPhone_X,
+      IPhone_8,
+      IPhone_7,
+      IPad_Pro_3,
+      IPad_7,
+      IPad_Air_2
+    ]
+  end
+end