+++ lib/applitools/selenium/visual_grid/ios_screen_orientation.rb
@@ -1,0 +1,14 @@
+module IosScreenshotOrientations
+  extend self
+  PORTRAIT ='portrait'
+  LANDSCAPE_LEFT = 'landscapeLeft'
+  LANDSCAPE_RIGHT = 'landscapeRight'
+
+  def enum_values
+    [
+      PORTRAIT,
+      LANDSCAPE_LEFT,
+      LANDSCAPE_RIGHT
+    ]
+  end
+end